fix: 他の型も同様に切り出し
