+++ src/types/ResponseMessage.ts
@@ -1,0 +1,10 @@
+import { ReactNode } from 'react'
+
+export type ResponseMessageType =
+  | {
+      status: 'success' | 'error'
+      text: ReactNode
+    }
+  | {
+      status: 'processing'
+    }
