+++ src/types/Size.ts
@@ -1,0 +1,1 @@
+export type Size = 's' | 'm' | 'l'
