+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export type { Gap, SeparateGap } from './Gap'
 export type { DecoratorsType, DecoratorType } from './Decorator'
-export type * from './props'
+export type { ResponseMessageType } from './ResponseMessage'
+export type { Size } from './Size'
